Move SchemaBuilder to root namespace
