+++ src/Error/formatError.php
@@ -5,17 +5,17 @@ namespace Digia\GraphQL\Error;
 use function Digia\GraphQL\Util\invariant;
 
 /**
- * @param GraphQLException $error
+ * @param GraphQLException|null $error
  * @return array
  * @throws InvariantException
  */
-function formatError(GraphQLException $error): array
+function formatError(?GraphQLException $error): array
 {
     invariant(null !== $error, 'Received null error.');
 
     return [
-        'message'       => $error->getMessage(),
-        'locations'     => $error->getLocationsAsArray(),
-        'path'          => $error->getPath(),
+        'message'   => $error->getMessage(),
+        'locations' => $error->getLocationsAsArray(),
+        'path'      => $error->getPath(),
     ];
 }
