+++ src/SchemaBuilder/DefinitionBuilderInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Digia\GraphQL\Language\SchemaBuilder;
+namespace Digia\GraphQL\SchemaBuilder;
 
 use Digia\GraphQL\Language\Node\DirectiveDefinitionNode;
 use Digia\GraphQL\Language\Node\NodeInterface;
