+++ src/SchemaBuilder/SchemaBuilderInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Digia\GraphQL\Language\SchemaBuilder;
+namespace Digia\GraphQL\SchemaBuilder;
 
 use Digia\GraphQL\Language\Node\DocumentNode;
 use Digia\GraphQL\Type\SchemaInterface;
