+++ src/api.php
@@ -12,8 +12,8 @@ use Digia\GraphQL\Language\Node\DocumentNode;
 use Digia\GraphQL\Language\Node\NodeInterface;
 use Digia\GraphQL\Language\ParserInterface;
 use Digia\GraphQL\Language\PrinterInterface;
-use Digia\GraphQL\Language\SchemaBuilder\SchemaBuilderInterface;
 use Digia\GraphQL\Language\Source;
+use Digia\GraphQL\SchemaBuilder\SchemaBuilderInterface;
 use Digia\GraphQL\SchemaValidator\SchemaValidatorInterface;
 use Digia\GraphQL\Type\SchemaInterface;
 use Digia\GraphQL\Util\SerializationInterface;
