--- tests/Functional/Language/schema-user-vote.graphqls
@@ -1,53 +1,0 @@
-schema {
-  query: Query
-  mutation: Mutation
-}
-
-type Query {
-  allLinks(filter: LinkFilter, skip: Int = 0, first: Int = 0): [Link]
-}
-
-type Mutation {
-  createLink(url: String!, description: String!): Link
-  createUser(name: String!, authProvider: AuthData!): User
-  signinUser(auth: AuthData): SigninPayload
-  createVote(linkId: ID, userId: ID): Vote
-}
-
-type Link {
-  id: ID!
-  url: String!
-  description: String
-  postedBy: User
-}
-
-type User {
-  id: ID!
-  name: String!
-  email: String
-  password: String
-}
-
-type SigninPayload {
-  token: String
-  user: User
-}
-
-type Vote {
-  id: ID!
-  createdAt: DateTime!
-  user: User!
-  link: Link!
-}
-
-input AuthData {
-  email: String!
-  password: String!
-}
-
-input LinkFilter {
-  description_contains: String
-  url_contains: String
-}
-
-scalar DateTime
