+++ tests/Functional/SchemaBuilder/SchemaBuilderTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Digia\GraphQL\Test\Functional\Language;
+namespace Digia\GraphQL\Test\Functional\SchemaBuilder;
 
 use Digia\GraphQL\Test\TestCase;
 use Digia\GraphQL\Type\Schema;
@@ -9,12 +9,11 @@ use function Digia\GraphQL\Util\readFile;
 
 class SchemaBuilderTest extends TestCase
 {
-
     public function testBuildsSchema()
     {
-        $introspectionQuery = readFile(__DIR__ . '/schema-user-vote.graphqls');
+        $source = readFile(__DIR__ . '/../starWars.graphqls');
 
-        $schema = buildSchema($introspectionQuery);
+        $schema = buildSchema($source);
 
         $this->assertInstanceOf(Schema::class, $schema);
     }
